+++ model/jpa/src/main/java/org/keycloak/models/jpa/entities/RealmAttributes.java
@@ -55,7 +55,6 @@ public interface RealmAttributes {
     String WEBAUTHN_POLICY_PASSKEYS_ENABLED = "webAuthnPolicyPasskeysEnabled";
 
     String ADMIN_EVENTS_EXPIRATION = "adminEventsExpiration";
-    String CLAIMS_SUPPORTED = "claimsSupported";
 
     String FIRST_BROKER_LOGIN_FLOW_ID = "firstBrokerLoginFlowId";
 
@@ -65,4 +64,5 @@ public interface RealmAttributes {
     String ADMIN_PERMISSIONS_ENABLED = "adminPermissionsEnabled";
     String ADMIN_PERMISSIONS_CLIENT_ID = "adminPermissionsClientId";
     String SCIM_ENABLED = "scimEnabled";
+    String ISSUER = "issuer";
 }