+++ services/src/main/java/org/keycloak/protocol/oidc/TokenManager.java
@@ -131,6 +131,7 @@ import org.keycloak.tracing.TracingAttributes;
 import org.keycloak.tracing.TracingProvider;
 import org.keycloak.util.JsonSerialization;
 import org.keycloak.util.TokenUtil;
+import org.keycloak.utils.StringUtil;
 
 import org.jboss.logging.Logger;
 
@@ -1384,7 +1385,7 @@ public class TokenManager {
             idToken.audience(client.getClientId());
             idToken.issuedNow();
             idToken.issuedFor(accessToken.getIssuedFor());
-            idToken.issuer(accessToken.getIssuer());
+            idToken.issuer(StringUtil.isNotBlank(realm.getIssuer()) && Boolean.valueOf(client.getAttribute(Constants.ISHARE_ENABLED)) ? realm.getIssuer() : accessToken.getIssuer());
             idToken.setNonce(clientSessionCtx.getAttribute(OIDCLoginProtocol.NONCE_PARAM, String.class));
             idToken.setSessionId(accessToken.getSessionId());
             idToken.exp(getIdTokenExpiration(accessToken.getExp()));