+++ services/src/test/java/org/keycloak/protocol/saml/SamlProtocolTest.java
@@ -665,6 +665,16 @@ public class SamlProtocolTest {
         }
 
         @Override
+        public String getIssuer() {
+            return null;
+        }
+
+        @Override
+        public void setIssuer(String issuer) {
+
+        }
+
+        @Override
         public List<String> getClaimsSupported() {
             return new ArrayList<>();
         }